Move classes used for generating Groovy interceptors from :core to :internal-instrumentation-api
So we can upgrade classes in core-api.
