+++ platforms/core-runtime/internal-instrumentation-api/src/main/java/org/gradle/internal/instrumentation/api/groovybytecode/Invocation.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.classpath.intercept;
-
-import org.codehaus.groovy.runtime.callsite.CallSite;
+package org.gradle.internal.instrumentation.api.groovybytecode;
 
 import javax.annotation.Nullable;
 
@@ -30,7 +28,7 @@ public interface Invocation {
      * It can be the instance if the invocation targets the instance method or property.
      *
      * @return the receiver of the method
-     * @see CallSite
+     * @see org.codehaus.groovy.runtime.callsite.CallSite
      */
     @Nullable
     Object getReceiver();
