+++ platforms/core-runtime/internal-instrumentation-api/src/main/java/org/gradle/internal/instrumentation/api/groovybytecode/InvocationUtils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.classpath.intercept;
+package org.gradle.internal.instrumentation.api.groovybytecode;
 
 import org.codehaus.groovy.runtime.wrappers.Wrapper;
 
 import javax.annotation.Nullable;
 
-class InvocationUtils {
+public class InvocationUtils {
     private InvocationUtils() {}
 
     @Nullable
-    static Object unwrap(Object obj) {
+    public static Object unwrap(Object obj) {
         if (obj instanceof Wrapper) {
             return ((Wrapper) obj).unwrap();
         }
