+++ platforms/core-runtime/internal-instrumentation-api/src/main/java/org/gradle/internal/instrumentation/api/groovybytecode/MethodHandleInvocation.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.classpath.intercept;
+package org.gradle.internal.instrumentation.api.groovybytecode;
 
 import javax.annotation.Nullable;
 import java.lang.invoke.MethodHandle;
 
-import static org.gradle.internal.classpath.intercept.InvocationUtils.unwrap;
+import static org.gradle.internal.instrumentation.api.groovybytecode.InvocationUtils.unwrap;
 
 /**
  * The implementation of {@link Invocation} that forwards the call to a MethodHandle. Supports both normal and spread Groovy calls.
