+++ platforms/core-runtime/internal-instrumentation-api/src/main/java/org/gradle/internal/instrumentation/api/groovybytecode/PropertyAwareCallInterceptor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.classpath.intercept;
+package org.gradle.internal.instrumentation.api.groovybytecode;
 
 import org.gradle.api.NonNullApi;
