+++ platforms/core-runtime/internal-instrumentation-api/src/main/java/org/gradle/internal/instrumentation/api/groovybytecode/package-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,10 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.classpath.intercept;
+/**
+ * Classes used for Groovy generated classes.
+ */
+@NonNullApi
+package org.gradle.internal.instrumentation.api.groovybytecode;
 
-import org.gradle.internal.instrumentation.api.groovybytecode.CallInterceptor;
-import org.gradle.internal.instrumentation.api.types.BytecodeInterceptorFilter;
-
-import java.util.List;
-
-public interface CallSiteInterceptorSet {
-    List<CallInterceptor> getCallInterceptors(BytecodeInterceptorFilter filter);
-}
+import org.gradle.api.NonNullApi;
