+++ subprojects/core/src/main/java/org/gradle/internal/classpath/CallInterceptingMetaClass.java
@@ -31,13 +31,13 @@ import org.codehaus.groovy.runtime.MetaClassHelper;
 import org.gradle.api.NonNullApi;
 import org.gradle.internal.Cast;
 import org.gradle.internal.Pair;
-import org.gradle.internal.classpath.intercept.CallInterceptor;
+import org.gradle.internal.instrumentation.api.groovybytecode.CallInterceptor;
 import org.gradle.internal.classpath.intercept.CallInterceptorResolver;
-import org.gradle.internal.classpath.intercept.InterceptScope;
-import org.gradle.internal.classpath.intercept.Invocation;
+import org.gradle.internal.instrumentation.api.groovybytecode.InterceptScope;
+import org.gradle.internal.instrumentation.api.groovybytecode.Invocation;
 import org.gradle.internal.classpath.intercept.InvocationImpl;
-import org.gradle.internal.classpath.intercept.PropertyAwareCallInterceptor;
-import org.gradle.internal.classpath.intercept.SignatureAwareCallInterceptor;
+import org.gradle.internal.instrumentation.api.groovybytecode.PropertyAwareCallInterceptor;
+import org.gradle.internal.instrumentation.api.groovybytecode.SignatureAwareCallInterceptor;
 import org.gradle.internal.metaobject.InstrumentedMetaClass;
 
 import javax.annotation.Nullable;
