+++ subprojects/core/src/main/java/org/gradle/internal/classpath/Instrumented.java
@@ -23,11 +23,11 @@ import org.codehaus.groovy.runtime.callsite.CallSite;
 import org.codehaus.groovy.runtime.callsite.CallSiteArray;
 import org.codehaus.groovy.vmplugin.v8.IndyInterface;
 import org.gradle.internal.SystemProperties;
-import org.gradle.internal.classpath.intercept.AbstractCallInterceptor;
+import org.gradle.internal.instrumentation.api.groovybytecode.AbstractCallInterceptor;
 import org.gradle.internal.classpath.intercept.ClassBoundCallInterceptor;
-import org.gradle.internal.classpath.intercept.FilterableCallInterceptor;
-import org.gradle.internal.classpath.intercept.InterceptScope;
-import org.gradle.internal.classpath.intercept.Invocation;
+import org.gradle.internal.instrumentation.api.groovybytecode.FilterableCallInterceptor;
+import org.gradle.internal.instrumentation.api.groovybytecode.InterceptScope;
+import org.gradle.internal.instrumentation.api.groovybytecode.Invocation;
 import org.gradle.internal.configuration.inputs.AccessTrackingEnvMap;
 import org.gradle.internal.configuration.inputs.AccessTrackingProperties;
 import org.gradle.internal.configuration.inputs.InstrumentedInputs;
